chore: create oscd-layout addon (#1510)
* chore: create oscd-layout addon...
- extracts all rendering UI logic including former Hosting mixin and Plugging mixin to the layout addon
- creates new events to connect layout addon to plugging loading functionality


* chore: fixes tests broken with refactor


* chore: adding missing <slot> tag to addon to allow composing


* chore: updating mock-open-scd to comply with new rendering logic


* chore: adding snapshot testing for layout


* chore: improving render functions' encapsulation and adding typedoc comments


* chore: include missing editCount prop on layout addon


* chore: modify import to get type declaration only


* chore: modify import to get type declaration only


* chore: modify import to get type declaration only


* chore: modify import to get type declaration only


* chore: modify import to get type declaration only


* chore: modify import to get type declaration only


* chore: modify import to get type declaration only


* chore: modify import to get type declaration only


* chore: add missing imports


* chore: import List class needed for instanceof validation


---------
